Basic Styling & Handler Refactor
- Basic style sheet now implemented. Makes site look a little nicer.

- Beginning to refactor existing handlers to use same standards as imported handlers.
