+++ src/NodeServer/public/html/liveweather.html
@@ -2,19 +2,23 @@
     <h2>Live Weather Data</h2>
 
     <section class="displayContent tempSection">
-        <p id="tempStat"></p>
+        <h3>Temperature</h3>
+        <p class="liveStat" id="tempStat"></p>
     </section>
 
     <section class="displayContent humidSection">
-        <p id="humidStat"></p>
+        <h3>Humidity</h3>
+        <p class="liveStat" id="humidStat"></p>
     </section>
 
     <section class="displayContent pressSection">
-        <p id="pressStat"></p>
+        <h3>Barometric Pressure</h3>
+        <p class="liveStat" id="pressStat"></p>
     </section>
 
     <section class="displayContent settingsSection">
         
     </section>
 
+    <script type="module" src="./js/updateStats.js" defer></script>
 </section>
