+++ src/NodeServer/public/index.html
@@ -3,6 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./css/style.css">
+    <link rel="stylesheet" href="./css/liveWeatherStyle.css">
     <title>Arduino Weather Dashboard</title>
 </head>
 <body>
@@ -15,15 +17,15 @@
             Weather History
         </button>
         <button class="navBtn" id="aboutApp" title="Information about This Project" data-page="/html/aboutapp.html">
-            About
+            About Project
         </button>
     </div>
-    <div id="displayContent">
+    <div class="displayContent">
         <h1>Welcome</h1>
         <p>This section is currently a work in progress. Check back in later to see updates!</p>
     </div>
-    <div id="footer">
-        <p>Gonnna need to add some licensing/legal stuff here</p>
+    <div class="footer">
+        <p>licensing/legal stuff here - Made by Jordan Locke 2026</p>
     </div>
 
     <script src="https://code.jquery.com/jquery-4.0.0.slim.min.js" integrity="sha256-8DGpv13HIm+5iDNWw1XqxgFB4mj+yOKFNb+tHBZOowc=" crossorigin="anonymous"></script>
