+++ src/NodeServer/server/handlers/weatherRoute.js
@@ -6,7 +6,7 @@
  * @todo
  * - Consider a better sorting system to take in more data, for now, could use multiple JSON files for different years, or simply move straight to a database
  */
-
+"use strict"
 const { collectRequestBody, appendWeatherFile } = require('../utils/requestParser');
 
 //Handles the routing for Arduino weather submissions
