+++ src/NodeServer/server/utils/resMsg.js
@@ -1,0 +1,9 @@
+//Small function to automate sending status messages to client
+"use strict"
+function resMsg(res, code, msg){
+    res.writeHead(code, {'Content-Type' : 'text/html'});
+    res.write(msg);
+    res.end();
+}
+
+exports.resMsg = resMsg;
